Separate MM optimization into module
Also, add a logistic rating conversion.

The goal of that scale is to produce human-readable ratings.
They are such that, assuming instantaneous performance follows a
Gumbel distribution (or more generically, assuming that rating
comparison follows a logistic distribution, which is widely studied
as a good statistical fit for observed results), a 10-point
difference indicates a 1% increase in winning rate.

The exact formula for the scale conversion is justified and
described in comments.